+++ MannikToolbox/MannikToolbox.csproj
@@ -135,10 +135,10 @@
     <Compile Include="Forms\ItemSearchForm.Designer.cs">
       <DependentUpon>ItemSearchForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="MainForm.cs">
+    <Compile Include="Forms\MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="MainForm.Designer.cs">
+    <Compile Include="Forms\MainForm.Designer.cs">
       <DependentUpon>MainForm.cs</DependentUpon>
     </Compile>
     <Compile Include="Forms\MobSearch.cs">
@@ -147,10 +147,10 @@
     <Compile Include="Forms\MobSearch.Designer.cs">
       <DependentUpon>MobSearch.cs</DependentUpon>
     </Compile>
-    <Compile Include="MySQLConfig.cs">
+    <Compile Include="Forms\MySQLConfig.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="MySQLConfig.Designer.cs">
+    <Compile Include="Forms\MySQLConfig.Designer.cs">
       <DependentUpon>MySQLConfig.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
@@ -192,14 +192,14 @@
     <EmbeddedResource Include="Forms\ItemSearchForm.resx">
       <DependentUpon>ItemSearchForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="MainForm.resx">
+    <EmbeddedResource Include="Forms\MainForm.resx">
       <DependentUpon>MainForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Forms\MobSearch.resx">
       <DependentUpon>MobSearch.cs</DependentUpon>
       <SubType>Designer</SubType>
     </EmbeddedResource>
-    <EmbeddedResource Include="MySQLConfig.resx">
+    <EmbeddedResource Include="Forms\MySQLConfig.resx">
       <DependentUpon>MySQLConfig.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
