+++ MannikToolbox/Program.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using System.Windows.Forms;
+using MannikToolbox.Forms;
 
 namespace MannikToolbox
 {
