+++ MannikToolbox/Services/ModelImageService.cs
@@ -1,4 +1,5 @@
-﻿using System.Drawing;
+﻿using System;
+using System.Drawing;
 using System.IO;
 using System.Net;
 using System.Reflection;
@@ -8,7 +9,7 @@ namespace MannikToolbox.Services
     public class ModelImageService
     {
         private static string BasePath => $@"{Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location)}\Assets\Models\";
-        
+
         public Image LoadMob(int modelId)
         {
             return LoadAsset(modelId, "mobs");
@@ -17,6 +18,17 @@ namespace MannikToolbox.Services
         public Image LoadItem(int modelId)
         {
             return LoadAsset(modelId, "items");
+        }
+        public Image LoadMob(int modelId, int maxWidth, int maxHeight)
+        {
+            var image = LoadMob(modelId);
+            return ScaleImage(image, maxWidth, maxHeight);
+        }
+
+        public Image LoadItem(int modelId, int maxWidth, int maxHeight)
+        {
+            var image = LoadItem(modelId);
+            return ScaleImage(image, maxWidth, maxHeight);
         }
 
         private Image LoadAsset(int modelId, string type)
@@ -52,6 +64,31 @@ namespace MannikToolbox.Services
             }
 
             return null;
+        }
+
+        private static Image ScaleImage(Image image, int maxWidth, int maxHeight)
+        {
+            if (image == null)
+            {
+                return null;
+            }
+
+            var ratioX = (double)maxWidth / image.Width;
+            var ratioY = (double)maxHeight / image.Height;
+            var ratio = Math.Min(ratioX, ratioY);
+
+            var newWidth = (int)(image.Width * ratio);
+            var newHeight = (int)(image.Height * ratio);
+
+            var newImage = new Bitmap(newWidth, newHeight);
+
+            using (var graphics = Graphics.FromImage(newImage))
+            {
+                graphics.DrawImage(image, 0, 0, newWidth, newHeight);
+            }
+            image.Dispose();
+
+            return newImage;
         }
     }
 }
